修复 xxl_job_log 主键id类型为 bigint(20) ，  防止 int(11) 位溢出导致调度任务不能正常调度。

+++ doc/db/tables_xxl_job.sql
@@ -30,7 +30,7 @@ CREATE TABLE `xxl_job_info` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `xxl_job_log` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
   `job_group` int(11) NOT NULL COMMENT '执行器主键ID',
   `job_id` int(11) NOT NULL COMMENT '任务，主键ID',
   `executor_address` varchar(255) DEFAULT NULL COMMENT '执行器地址，本次执行的地址',
