+++ xxl-job-core/src/main/java/com/xxl/job/core/log/XxlJobFileAppender.java
@@ -68,7 +68,7 @@ public class XxlJobFileAppender {
 	 * @param logId
 	 * @return
 	 */
-	public static String makeLogFileName(Date triggerDate, int logId) {
+	public static String makeLogFileName(Date triggerDate, long logId) {
 
 		// filePath/yyyy-MM-dd
 		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");	// avoid concurrent problem, can not be static
